python_backend/requirements.txt
loguru
requests
python-dotenv
pydantic
pydantic-settings
openai
tenacity
jinja2
weasyprint
supabase
matplotlib
duckduckgo-search
trafilatura
lxml_html_clean
